Fix a time resolution bug

+++ Daigassou/Input_Midi/MidiToKey.cs
@@ -287,7 +287,7 @@ namespace Daigassou
                             break;
                         default:
                         {
-                            nowTimeMs += (int) (tickBase * ev.Event.DeltaTime);
+                            nowTimeMs += (tickBase * ev.Event.DeltaTime);
                         }
                             break;
                     }
